Modified animate function, made it more compact using vpython module.
